Add eclipse projects, fixes

+++ soft/WiFiAPP/src/ffs.c
@@ -54,17 +54,17 @@ static void f_init(void) {
 }
 
 static void f_read(uint32_t pos, uint8_t *data, int size) {
-	ets_printf("f_read(0x%05X, %d)\n", pos, size);
+	ets_printf("f_read(0x%05X, %d)\n", (unsigned)pos, size);
 	spi_flash_read(FFS_AT + pos, (uint32*) data, size);
 }
 
 static void f_write(uint32_t pos, const uint8_t *data, int size) {
-	ets_printf("f_write(0x%05X, %d)\n", pos, size);
+	ets_printf("f_write(0x%05X, %d)\n", (unsigned)pos, size);
 	spi_flash_write(FFS_AT + pos, (const uint32*) data, size);
 }
 
 static void f_erase(uint32_t pos) {
-	ets_printf("f_erase(0x%05X)\n", pos);
+	ets_printf("f_erase(0x%05X)\n", (unsigned)pos);
 	spi_flash_erase_sector((FFS_AT + pos) / 4096);
 }
